requirements.txt
aiohttp==3.7.4.post0
beautifulsoup4==4.10.0
discord.py==1.7.3
giphy_client==1.0.0
hentai==3.2.9
libgen_api==0.5.0
loguru==0.5.3
lxml==4.6.4
lyricsgenius==3.0.1
pornhub_api==0.2.0
pornhubapi==0.1.0
requests==2.26.0
validators==0.18.2
wavelink==0.9.10
wikipedia==1.4.0
Pillow==8.4.0
editdistance==0.6.0